+++ src/main/java/ua/nure/danielost/shuttler/service/VehicleServiceImpl.java
@@ -2,16 +2,16 @@ package ua.nure.danielost.shuttler.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import ua.nure.danielost.shuttler.configuration.SecurityConfig;
 import ua.nure.danielost.shuttler.exception.EmptyTableException;
 import ua.nure.danielost.shuttler.exception.InvalidIdException;
-import ua.nure.danielost.shuttler.exception.NoSuchRouteException;
 import ua.nure.danielost.shuttler.exception.NoSuchVehicleException;
 import ua.nure.danielost.shuttler.model.Route;
 import ua.nure.danielost.shuttler.model.User;
 import ua.nure.danielost.shuttler.model.Vehicle;
-import ua.nure.danielost.shuttler.repository.RouteRepository;
-import ua.nure.danielost.shuttler.repository.UserRepository;
-import ua.nure.danielost.shuttler.repository.VehicleRepository;
+import ua.nure.danielost.shuttler.security.jwt.repository.RouteRepository;
+import ua.nure.danielost.shuttler.security.jwt.repository.UserRepository;
+import ua.nure.danielost.shuttler.security.jwt.repository.VehicleRepository;
 
 import java.util.List;
 import java.util.Optional;
@@ -38,11 +38,17 @@ public class VehicleServiceImpl implements VehicleService {
     }
 
     @Override
-    public void deleteVehicle(String vin) throws NoSuchVehicleException {
+    public void deleteVehicle(String vin) throws NoSuchVehicleException, InvalidIdException {
         Optional<Vehicle> vehicleOptional = vehicleRepository.findById(vin);
         if (!vehicleOptional.isPresent()) {
             throw new NoSuchVehicleException("No vehicles with " + vin + " vin code in database");
         }
+        String username = SecurityConfig.getUsernameByContext();
+        Vehicle vehicle = vehicleOptional.get();
+        if (!vehicle.getCreator().getUsername().equals(username)) {
+            throw new InvalidIdException("You can delete only your own vehicles");
+        }
+
         vehicleRepository.delete(vehicleOptional.get());
     }
 
@@ -85,7 +91,7 @@ public class VehicleServiceImpl implements VehicleService {
             Vehicle vehicle,
             String vin,
             long route_id
-    ) throws NoSuchVehicleException {
+    ) throws NoSuchVehicleException, InvalidIdException {
         Optional<Vehicle> vehicleOptional = vehicleRepository.findById(vin);
         if (!vehicleOptional.isPresent()) {
             throw new NoSuchVehicleException("No vehicles with " + vin + " vin code in database");
@@ -97,7 +103,15 @@ public class VehicleServiceImpl implements VehicleService {
         }
 
         Vehicle vehicleToUpdate = vehicleOptional.get();
-        vehicleToUpdate.setRoute(routeOptional.get());
+        Route route = routeOptional.get();
+
+        String username = SecurityConfig.getUsernameByContext();
+        if (!vehicleToUpdate.getCreator().getUsername().equals(username) ||
+            !route.getCreator().getUsername().equals(username)) {
+            throw new InvalidIdException("You can update only your vehicle and only your routes");
+        }
+
+        vehicleToUpdate.setRoute(route);
         vehicleToUpdate.setCurrent_capacity(vehicle.getCurrent_capacity());
         vehicleToUpdate.setMax_capacity(vehicle.getMax_capacity());
 
@@ -105,13 +119,29 @@ public class VehicleServiceImpl implements VehicleService {
     }
 
     @Override
-    public Vehicle addVehicle(Vehicle vehicle, long routeId, long userId) throws InvalidIdException {
+    public Vehicle addVehicle(
+            Vehicle vehicle,
+            long routeId,
+            long userId
+    ) throws InvalidIdException {
         Optional<Route> routeOptional = routeRepository.findById(routeId);
         Optional<User> userOptional = userRepository.findById(userId);
+        Optional<Vehicle> vehicleOptional = vehicleRepository.findById(vehicle.getVin());
         if (!routeOptional.isPresent() || !userOptional.isPresent()) {
             throw new InvalidIdException("Invalid user or route ID");
         }
-        vehicle.setRoute(routeOptional.get());
+
+        if (vehicleOptional.isPresent()) {
+            throw new InvalidIdException("Vehicle with this VIN {" + vehicle.getVin() + "} exists");
+        }
+
+        Route route = routeOptional.get();
+        String username = SecurityConfig.getUsernameByContext();
+        if (!route.getCreator().getUsername().equals(username)) {
+            throw new InvalidIdException("You can add vehicles only to your routes");
+        }
+
+        vehicle.setRoute(route);
         vehicle.setCreator(userOptional.get());
         return vehicleRepository.save(vehicle);
     }
